+++ flybywire-aircraft-a320-neo/html_ui/Pages/VCockpit/Instruments/Airliners/FlyByWire_A320_Neo/CDU/A320_Neo_CDU.html
@@ -161,7 +161,7 @@
 <script type="text/html" import-script="/Pages/A32NX_Core/A32NX_Core.js"></script>
 <script type="text/html" import-script="/Pages/A32NX_Core/A32NX_ATSU.js"></script>
 <script type="text/html" import-script="/Pages/A32NX_Core/A32NX_FuelPred.js"></script>
-<script type="text/html" import-script="/Pages/A32NX_Core/A32NX_ROPS.js"></script>
+<script type="text/html" import-script="/Pages/A32NX_Core/FMGC/A32NX_ROPS.js"></script>
 <script type="text/html" import-script="/Pages/A32NX_Core/FMGC/A32NX_ConstraintManager.js"></script>
 <script type="text/html" import-script="/Pages/A32NX_Core/FMGC/A32NX_FlightPhaseManager.js"></script>
