+++ flybywire-aircraft-a320-neo/html_ui/Pages/VCockpit/Instruments/Airliners/FlyByWire_A320_Neo/FMC/A32NX_FMCMainDisplay.js
@@ -483,7 +483,7 @@ class FMCMainDisplay extends BaseAirliners {
             }
         }, 15000);
 
-        this.rops = new A32NX_ROPS(this);
+        this.rops = new Fmgc.Rops(this);
         this.rops.init();
     }
