+++ src/fmgc/src/index.ts
@@ -3,6 +3,8 @@ import { FlightPlanAsoboSync } from './flightplanning/FlightPlanAsoboSync';
 import { GuidanceManager } from './guidance/GuidanceManager';
 import { ManagedFlightPlan } from './flightplanning/ManagedFlightPlan';
 import { GuidanceController } from './guidance/GuidanceController';
+import { NavRadioManager } from './radionav/NavRadioManager';
+import { ROPS as Rops } from "./rops/ROPS";
 
 export {
     FlightPlanManager,
@@ -10,4 +12,6 @@ export {
     FlightPlanAsoboSync,
     GuidanceManager,
     GuidanceController,
+    NavRadioManager,
+    Rops,
 };
