+++ typings/fs-base-ui/html_ui/JS/simvar.d.ts
@@ -111,6 +111,4 @@ declare global {
     }
 }
 
-export {
-    SimVar
-}
+export {};
